@@ -348,7 +348,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         /// <remarks>
         /// See SymbolPreparer::checkIfaceHiding.
         /// </remarks>
-        private static OverriddenOrHiddenMembersResult MakeInterfaceOverriddenOrHiddenMembers(Symbol member, bool memberIsFromSomeCompilation)
+        internal static OverriddenOrHiddenMembersResult MakeInterfaceOverriddenOrHiddenMembers(Symbol member, bool memberIsFromSomeCompilation)
         {
             Debug.Assert(!member.IsAccessor());
 
@@ -518,84 +518,111 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                     int otherMemberArity = otherMember.GetMemberArity();
                     if (otherMemberArity == memberArity || (memberKind == SymbolKind.Method && otherMemberArity == 0))
                     {
-                        AddHiddenMemberIfApplicable(ref hiddenBuilder, member.Kind, otherMember);
+                        AddHiddenMemberIfApplicable(ref hiddenBuilder, memberKind, otherMember);
                     }
                 }
                 else if (!currTypeHasExactMatch)
                 {
-                    if (exactMatchComparer.Equals(member, otherMember))
+                    switch (memberKind)
                     {
-                        currTypeHasExactMatch = true;
-                        currTypeBestMatch = otherMember;
-                    }
-                    else if (fallbackComparer.Equals(member, otherMember))
-                    {
-                        // If this method is from source, we'll also consider methods that match
-                        // without regard to custom modifiers.  If there's more than one, we'll
-                        // choose the one with the fewest custom modifiers.
-                        int methodCustomModifierCount = CustomModifierCount(otherMember);
-                        if (methodCustomModifierCount < minCustomModifierCount)
-                        {
-                            Debug.Assert(memberIsFromSomeCompilation || minCustomModifierCount == int.MaxValue, "Metadata members require exact custom modifier matches.");
-                            minCustomModifierCount = methodCustomModifierCount;
+                        case SymbolKind.Field:
+                            currTypeHasExactMatch = true;
                             currTypeBestMatch = otherMember;
-                        }
-                    }
-                    else
-                    {
-                        currTypeHasSameKindNonMatch = true;
+                            break;
+                        case SymbolKind.NamedType:
+                            if (otherMember.GetMemberArity() == memberArity)
+                            {
+                                currTypeHasExactMatch = true;
+                                currTypeBestMatch = otherMember;
+                            }
+                            break;
+
+                        default:
+                            if (exactMatchComparer.Equals(member, otherMember))
+                            {
+                                currTypeHasExactMatch = true;
+                                currTypeBestMatch = otherMember;
+                            }
+                            else if (fallbackComparer.Equals(member, otherMember))
+                            {
+                                // If this method is from source, we'll also consider methods that match
+                                // without regard to custom modifiers.  If there's more than one, we'll
+                                // choose the one with the fewest custom modifiers.
+                                int methodCustomModifierCount = CustomModifierCount(otherMember);
+                                if (methodCustomModifierCount < minCustomModifierCount)
+                                {
+                                    Debug.Assert(memberIsFromSomeCompilation || minCustomModifierCount == int.MaxValue, "Metadata members require exact custom modifier matches.");
+                                    minCustomModifierCount = methodCustomModifierCount;
+                                    currTypeBestMatch = otherMember;
+                                }
+                            }
+                            else
+                            {
+                                currTypeHasSameKindNonMatch = true;
+                            }
+
+                            break;
                     }
                 }
             }
 
-            // If the member is from metadata, then even a fallback match is an exact match.
-            // We just declined to set the flag at the time in case there was a "better" exact match.
-            // Having said that, there's no reason to update the flag, since no-one will consume it.
-            // if (!memberIsFromSomeCompilation && ((object)currTypeBestMatch != null)) currTypeHasExactMatch = true;
-
-            // There's a special case where we have to go back and fix up our best match.
-            // If member is from source, then it has no custom modifiers (at least,
-            // until it copies them from the member it overrides, which we're trying to
-            // compute now).  Therefore, an exact match will be one that has no custom
-            // modifiers in/on its parameters.  The best match may, however, have custom
-            // modifiers in/on its (return) type, since that wasn't considered during the
-            // signature comparison.  If that is the case, then we need to make sure that
-            // there isn't another inexact match (i.e. same signature ignoring custom 
-            // modifiers) with fewer custom modifiers.
-            // NOTE: If member is constructed, then it has already inherited custom modifiers
-            // from the underlying member and this cleanup is unnecessary.  That's why we're
-            // checking member.IsDefinition in addition to memberIsFromSomeCompilation.
-            if (currTypeHasExactMatch && memberIsFromSomeCompilation && member.IsDefinition && TypeOrReturnTypeHasCustomModifiers(currTypeBestMatch))
+            switch (memberKind)
             {
+                case SymbolKind.Field:
+                case SymbolKind.NamedType:
+                    break;
+
+                default:
+                    // If the member is from metadata, then even a fallback match is an exact match.
+                    // We just declined to set the flag at the time in case there was a "better" exact match.
+                    // Having said that, there's no reason to update the flag, since no-one will consume it.
+                    // if (!memberIsFromSomeCompilation && ((object)currTypeBestMatch != null)) currTypeHasExactMatch = true;
+
+                    // There's a special case where we have to go back and fix up our best match.
+                    // If member is from source, then it has no custom modifiers (at least,
+                    // until it copies them from the member it overrides, which we're trying to
+                    // compute now).  Therefore, an exact match will be one that has no custom
+                    // modifiers in/on its parameters.  The best match may, however, have custom
+                    // modifiers in/on its (return) type, since that wasn't considered during the
+                    // signature comparison.  If that is the case, then we need to make sure that
+                    // there isn't another inexact match (i.e. same signature ignoring custom 
+                    // modifiers) with fewer custom modifiers.
+                    // NOTE: If member is constructed, then it has already inherited custom modifiers
+                    // from the underlying member and this cleanup is unnecessary.  That's why we're
+                    // checking member.IsDefinition in addition to memberIsFromSomeCompilation.
+                    if (currTypeHasExactMatch && memberIsFromSomeCompilation && member.IsDefinition && TypeOrReturnTypeHasCustomModifiers(currTypeBestMatch))
+                    {
 #if DEBUG
-                // If there were custom modifiers on the parameters, then the match wouldn't have been
-                // exact and so we would already have applied the custom modifier count as a tie-breaker.
-                foreach (ParameterSymbol param in currTypeBestMatch.GetParameters())
-                {
-                    Debug.Assert(!(param.TypeWithAnnotations.CustomModifiers.Any() || param.RefCustomModifiers.Any()));
-                    Debug.Assert(!param.Type.HasCustomModifiers(flagNonDefaultArraySizesOrLowerBounds: false));
-                }
+                        // If there were custom modifiers on the parameters, then the match wouldn't have been
+                        // exact and so we would already have applied the custom modifier count as a tie-breaker.
+                        foreach (ParameterSymbol param in currTypeBestMatch.GetParameters())
+                        {
+                            Debug.Assert(!(param.TypeWithAnnotations.CustomModifiers.Any() || param.RefCustomModifiers.Any()));
+                            Debug.Assert(!param.Type.HasCustomModifiers(flagNonDefaultArraySizesOrLowerBounds: false));
+                        }
 #endif
 
-                Symbol minCustomModifierMatch = currTypeBestMatch;
+                        Symbol minCustomModifierMatch = currTypeBestMatch;
 
-                foreach (Symbol otherMember in currType.GetMembers(member.Name))
-                {
-                    if (otherMember.Kind == currTypeBestMatch.Kind && !ReferenceEquals(otherMember, currTypeBestMatch))
-                    {
-                        if (MemberSignatureComparer.CSharpOverrideComparer.Equals(otherMember, currTypeBestMatch))
+                        foreach (Symbol otherMember in currType.GetMembers(member.Name))
                         {
-                            int customModifierCount = CustomModifierCount(otherMember);
-                            if (customModifierCount < minCustomModifierCount)
+                            if (otherMember.Kind == currTypeBestMatch.Kind && !ReferenceEquals(otherMember, currTypeBestMatch))
                             {
-                                minCustomModifierCount = customModifierCount;
-                                minCustomModifierMatch = otherMember;
+                                if (MemberSignatureComparer.CSharpOverrideComparer.Equals(otherMember, currTypeBestMatch))
+                                {
+                                    int customModifierCount = CustomModifierCount(otherMember);
+                                    if (customModifierCount < minCustomModifierCount)
+                                    {
+                                        minCustomModifierCount = customModifierCount;
+                                        minCustomModifierMatch = otherMember;
+                                    }
+                                }
                             }
                         }
-                    }
-                }
 
-                currTypeBestMatch = minCustomModifierMatch;
+                        currTypeBestMatch = minCustomModifierMatch;
+                    }
+                    break;
             }
         }
 
@@ -619,7 +646,8 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
 
             if ((object)representativeMember != null)
             {
-                bool needToSearchForRelated = !representativeMember.ContainingType.IsDefinition || representativeMember.IsIndexer();
+                bool needToSearchForRelated = representativeMember.Kind != SymbolKind.Field && representativeMember.Kind != SymbolKind.NamedType &&
+                                              (!representativeMember.ContainingType.IsDefinition || representativeMember.IsIndexer());
 
                 if (isOverride)
                 {
@@ -788,6 +816,8 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         private static void FindOtherHiddenMembersInContainingType(SymbolKind hidingMemberKind, Symbol representativeMember, ref ArrayBuilder<Symbol> hiddenBuilder)
         {
             Debug.Assert((object)representativeMember != null);
+            Debug.Assert(representativeMember.Kind != SymbolKind.Field);
+            Debug.Assert(representativeMember.Kind != SymbolKind.NamedType);
             Debug.Assert(representativeMember.Kind == SymbolKind.Property || !representativeMember.ContainingType.IsDefinition);
 
             IEqualityComparer<Symbol> comparer = MemberSignatureComparer.CSharpCustomModifierOverrideComparer;
