@@ -375,7 +375,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         protected virtual ImmutableArray<NamedTypeSymbol> MakeAllInterfaces()
         {
             var result = ArrayBuilder<NamedTypeSymbol>.GetInstance();
-            var visited = new HashSet<NamedTypeSymbol>(SymbolEqualityComparer.ConsiderEverything);
+            var visited = new HashSet<NamedTypeSymbol>(TypeSymbol.EqualsConsiderEverything);
 
             for (var baseType = this; !ReferenceEquals(baseType, null); baseType = baseType.BaseTypeNoUseSiteDiagnostics)
             {
@@ -477,7 +477,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         /// <param name="interfaceMember">
         /// Must be a non-null interface property, method, or event.
         /// </param>
-        public Symbol FindImplementationForInterfaceMember(Symbol interfaceMember, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        public Symbol FindImplementationForInterfaceMember(Symbol interfaceMember)
         {
             if ((object)interfaceMember == null)
             {
@@ -491,10 +491,11 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
 
             if (this.IsInterfaceType())
             {
+                HashSet<DiagnosticInfo> useSiteDiagnostics = null;
                 return FindMostSpecificImplementation(interfaceMember, (NamedTypeSymbol)this, ref useSiteDiagnostics);
             }
 
-            return FindImplementationForInterfaceMemberInNonInterfaceWithDiagnostics(interfaceMember).Symbol;
+            return FindImplementationForInterfaceMemberInNonInterface(interfaceMember);
         }
 
         /// <summary>
@@ -769,9 +770,8 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
 
         ISymbol ITypeSymbol.FindImplementationForInterfaceMember(ISymbol interfaceMember)
         {
-            HashSet<DiagnosticInfo> useSiteDiagnostics = null;
-            return interfaceMember is Symbol
-                ? FindImplementationForInterfaceMember((Symbol)interfaceMember, ref useSiteDiagnostics)
+            return interfaceMember is Symbol symbol
+                ? FindImplementationForInterfaceMember(symbol)
                 : null;
         }
 
@@ -812,7 +812,34 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
 
         #region Interface member checks
 
-        protected SymbolAndDiagnostics FindImplementationForInterfaceMemberInNonInterfaceWithDiagnostics(Symbol interfaceMember)
+        /// <summary>
+        /// Locate implementation of the <paramref name="interfaceMember"/> in context of the current type.
+        /// The method is using cache to optimize subsequent calls for the same <paramref name="interfaceMember"/>.
+        /// </summary>
+        /// <param name="interfaceMember">Member for which an implementation should be found.</param>
+        /// <param name="ignoreImplementationInInterfacesIfResultIsNotReady">
+        /// The process of looking up an implementation for an accessor can involve figuring out how corresponding event/property is implemented,
+        /// <see cref="CheckForImplementationOfCorrespondingPropertyOrEvent"/>. And the process of looking up an implementation for a property can
+        /// involve figuring out how corresponding accessors are implemented, <see cref="FindMostSpecificImplementationInInterfaces"/>. This can 
+        /// lead to cycles, which could be avoided if we ignore the presence of implementations in interfaces for the purpose of
+        /// <see cref="CheckForImplementationOfCorrespondingPropertyOrEvent"/>. Fortunately, logic in it allows us to ignore the presence of
+        /// implementations in interfaces and we use that.
+        /// When the value of this parameter is true and the result that takes presence of implementations in interfaces into account is not
+        /// available from the cache, the lookup will be performed ignoring the presence of implementations in interfaces. Otherwise, result from
+        /// the cache is returned.
+        /// When the value of the parameter is false, the result from the cache is returned, or calculated, taking presence of implementations
+        /// in interfaces into account and then cached.
+        /// This means that:
+        ///  - A symbol from an interface can still be returned even when <paramref name="ignoreImplementationInInterfacesIfResultIsNotReady"/> is true.
+        ///    A subsequent call with <paramref name="ignoreImplementationInInterfacesIfResultIsNotReady"/> false will return the same value. 
+        ///  - If symbol from a non-interface is returned when <paramref name="ignoreImplementationInInterfacesIfResultIsNotReady"/> is true. A subsequent
+        ///    call with <paramref name="ignoreImplementationInInterfacesIfResultIsNotReady"/> false will return the same value.
+        ///  - If no symbol is returned for <paramref name="ignoreImplementationInInterfacesIfResultIsNotReady"/> true. A subsequent call with
+        ///    <paramref name="ignoreImplementationInInterfacesIfResultIsNotReady"/> might return a symbol, but that symbol guaranteed to be from an interface.
+        ///  - If the first request is done with <paramref name="ignoreImplementationInInterfacesIfResultIsNotReady"/> false. A subsequent call
+        ///    is guaranted to return the same result regardless of <paramref name="ignoreImplementationInInterfacesIfResultIsNotReady"/> value.
+        /// </param>
+        protected SymbolAndDiagnostics FindImplementationForInterfaceMemberInNonInterfaceWithDiagnostics(Symbol interfaceMember, bool ignoreImplementationInInterfacesIfResultIsNotReady = false)
         {
             Debug.Assert((object)interfaceMember != null);
             Debug.Assert(!this.IsInterfaceType());
@@ -847,8 +874,15 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                         return result;
                     }
 
-                    result = ComputeImplementationAndDiagnosticsForInterfaceMember(interfaceMember);
-                    map.TryAdd(interfaceMember, result);
+                    result = ComputeImplementationAndDiagnosticsForInterfaceMember(interfaceMember, ignoreImplementationInInterfaces: ignoreImplementationInInterfacesIfResultIsNotReady,
+                                                                                   out bool implementationInInterfacesMightChangeResult);
+
+                    Debug.Assert(ignoreImplementationInInterfacesIfResultIsNotReady || !implementationInInterfacesMightChangeResult);
+                    Debug.Assert(!implementationInInterfacesMightChangeResult || result.Symbol is null); ;
+                    if (!implementationInInterfacesMightChangeResult)
+                    {
+                        map.TryAdd(interfaceMember, result);
+                    }
                     return result;
 
                 default:
@@ -856,10 +890,15 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             }
         }
 
-        private SymbolAndDiagnostics ComputeImplementationAndDiagnosticsForInterfaceMember(Symbol interfaceMember)
+        internal Symbol FindImplementationForInterfaceMemberInNonInterface(Symbol interfaceMember, bool ignoreImplementationInInterfacesIfResultIsNotReady = false)
+        {
+            return FindImplementationForInterfaceMemberInNonInterfaceWithDiagnostics(interfaceMember, ignoreImplementationInInterfacesIfResultIsNotReady).Symbol;
+        }
+
+        private SymbolAndDiagnostics ComputeImplementationAndDiagnosticsForInterfaceMember(Symbol interfaceMember, bool ignoreImplementationInInterfaces, out bool implementationInInterfacesMightChangeResult)
         {
             var diagnostics = DiagnosticBag.GetInstance();
-            var implementingMember = ComputeImplementationForInterfaceMember(interfaceMember, this, diagnostics);
+            var implementingMember = ComputeImplementationForInterfaceMember(interfaceMember, this, diagnostics, ignoreImplementationInInterfaces, out implementationInInterfacesMightChangeResult);
             var implementingMemberAndDiagnostics = new SymbolAndDiagnostics(implementingMember, diagnostics.ToReadOnlyAndFree());
             return implementingMemberAndDiagnostics;
         }
@@ -873,8 +912,17 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         /// <param name="interfaceMember">A non-null implementable member on an interface type.</param>
         /// <param name="implementingType">The type implementing the interface property (usually "this").</param>
         /// <param name="diagnostics">Bag to which to add diagnostics.</param>
+        /// <param name="ignoreImplementationInInterfaces">Do not consider implementation in an interface as a valid candidate for the purpose of this computation.</param>
+        /// <param name="implementationInInterfacesMightChangeResult">
+        /// Returns true when <paramref name="ignoreImplementationInInterfaces"/> is true, the method fails to locate an implementation and an implementation in
+        /// an interface, if any (its presence is not checked), could potentially be a candidate. Returns false otherwise.
+        /// When true is returned, a different call with <paramref name="ignoreImplementationInInterfaces"/> false might return a symbol. That symbol, if any,
+        /// is guaranteed to be from an interface.
+        /// This parameter is used to optimize caching in <see cref="FindImplementationForInterfaceMemberInNonInterfaceWithDiagnostics"/>.
+        /// </param>
         /// <returns>The implementing property or null, if there isn't one.</returns>
-        private static Symbol ComputeImplementationForInterfaceMember(Symbol interfaceMember, TypeSymbol implementingType, DiagnosticBag diagnostics)
+        private static Symbol ComputeImplementationForInterfaceMember(Symbol interfaceMember, TypeSymbol implementingType, DiagnosticBag diagnostics,
+                                                                      bool ignoreImplementationInInterfaces, out bool implementationInInterfacesMightChangeResult)
         {
             Debug.Assert(!implementingType.IsInterfaceType());
             Debug.Assert(interfaceMember.Kind == SymbolKind.Method || interfaceMember.Kind == SymbolKind.Property || interfaceMember.Kind == SymbolKind.Event);
@@ -918,19 +966,22 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                 MultiDictionary<Symbol, Symbol>.ValueSet explicitImpl = currType.GetExplicitImplementationForInterfaceMember(interfaceMember);
                 if (explicitImpl.Count == 1)
                 {
+                    implementationInInterfacesMightChangeResult = false;
                     return explicitImpl.Single();
                 }
                 else if (explicitImpl.Count > 1)
                 {
                     diagnostics.Add(ErrorCode.ERR_DuplicateExplicitImpl, implementingType.Locations[0], interfaceMember);
+                    implementationInInterfacesMightChangeResult = false;
                     return null;
                 }
 
-                // WORKAROUND: see comment on method.
-                // https://github.com/dotnet/roslyn/issues/34452: Is this workaround still relevant? Do we need it when we are looking
-                //                                                for implementations in interfaces?
                 if (IsExplicitlyImplementedViaAccessors(interfaceMember, currType, out Symbol currTypeExplicitImpl))
                 {
+                    // We are looking for a property or event implementation and found an explicit implementation
+                    // for its accessor(s) in this type. Stop the process and return event/property associated
+                    // with the accessor(s), if any.
+                    implementationInInterfacesMightChangeResult = false;
                     // NOTE: may be null.
                     return currTypeExplicitImpl;
                 }
@@ -948,7 +999,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                         }
                     }
                 }
-
 
                 // We want the implementation from the most derived type at or above the first one to
                 // include the interface (or a subinterface) in its interface list
@@ -980,19 +1030,41 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
 
             Debug.Assert(!canBeImplementedImplicitly || (object)implementingBaseOpt == null);
 
+            bool tryDefaultInterfaceImplementation = true;
+
             // Dev10 has some extra restrictions and extra wiggle room when finding implicit
             // implementations for interface accessors.  Perform some extra checks and possibly
             // update the result (i.e. implicitImpl).
             if (interfaceMember.IsAccessor())
             {
-                // https://github.com/dotnet/roslyn/issues/34453: Do we need to adjust behavior of this function in any way?
-                CheckForImplementationOfCorrespondingPropertyOrEvent((MethodSymbol)interfaceMember, implementingType, implementingTypeIsFromSomeCompilation, ref implicitImpl, ref useSiteDiagnostics);
+                Symbol originalImplicitImpl = implicitImpl;
+                CheckForImplementationOfCorrespondingPropertyOrEvent((MethodSymbol)interfaceMember, implementingType, implementingTypeIsFromSomeCompilation, ref implicitImpl);
+
+                // If we discarded the candidate, we don't want default interface implementation to take over later, since runtime might still use the discarded candidate.
+                if (originalImplicitImpl is object && implicitImpl is null)
+                {
+                    tryDefaultInterfaceImplementation = false;
+                }
             }
 
-            if ((object)implicitImpl == null && seenTypeDeclaringInterface)
+            Symbol defaultImpl = null;
+
+            if ((object)implicitImpl == null && seenTypeDeclaringInterface && tryDefaultInterfaceImplementation)
             {
-                // Check for default interface implementations
-                implicitImpl = FindMostSpecificImplementationInInterfaces(interfaceMember, implementingType, ref useSiteDiagnostics, diagnostics);
+                if (ignoreImplementationInInterfaces)
+                {
+                    implementationInInterfacesMightChangeResult = true;
+                }
+                else
+                {
+                    // Check for default interface implementations
+                    defaultImpl = FindMostSpecificImplementationInInterfaces(interfaceMember, implementingType, ref useSiteDiagnostics, diagnostics);
+                    implementationInInterfacesMightChangeResult = false;
+                }
+            }
+            else
+            {
+                implementationInInterfacesMightChangeResult = false;
             }
 
 #if !DEBUG
@@ -1003,9 +1075,15 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                 diagnostics.Add(GetInterfaceLocation(interfaceMember, implementingType), useSiteDiagnostics);
             }
 
+            if (defaultImpl is object)
+            {
+                ReportDefaultInterfaceImplementationMatchDiagnostics(interfaceMember, implementingType, defaultImpl, diagnostics);
+                return defaultImpl;
+            }
+
             if ((object)implicitImpl != null)
             {
-                if (!canBeImplementedImplicitly && !implicitImpl.ContainingType.IsInterface)
+                if (!canBeImplementedImplicitly)
                 {
                     if (interfaceMember.Kind == SymbolKind.Method &&
                         (object)implementingBaseOpt == null) // Otherwise any approprite errors are going to be reported for the base.
@@ -1033,13 +1111,24 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                                                                          ref HashSet<DiagnosticInfo> useSiteDiagnostics,
                                                                          DiagnosticBag diagnostics)
         {
-            Symbol implicitImpl = FindMostSpecificImplementationInBases(interfaceMember,
-                                                                        implementingType,
-                                                                        ref useSiteDiagnostics, out Symbol conflict1, out Symbol conflict2);
+            Debug.Assert(!implementingType.IsInterfaceType());
+
+            // If we are dealing with a property or event and an implementation of at least one accessor is not from an interface, it 
+            // wouldn't be right to say that the event/property is implemented in an interface because its accessor isn't.
+            (MethodSymbol interfaceAccessor1, MethodSymbol interfaceAccessor2) = GetImplementableAccessors(interfaceMember);
+
+            if (stopLookup(interfaceAccessor1, implementingType) || stopLookup(interfaceAccessor2, implementingType))
+            {
+                return null;
+            }
+
+            Symbol defaultImpl = FindMostSpecificImplementationInBases(interfaceMember,
+                                                                       implementingType,
+                                                                       ref useSiteDiagnostics, out Symbol conflict1, out Symbol conflict2);
 
             if ((object)conflict1 != null)
             {
-                Debug.Assert((object)implicitImpl == null);
+                Debug.Assert((object)defaultImpl == null);
                 Debug.Assert((object)conflict2 != null);
                 diagnostics.Add(ErrorCode.ERR_MostSpecificImplementationIsNotFound, GetInterfaceLocation(interfaceMember, implementingType),
                                 interfaceMember, conflict1, conflict2);
@@ -1047,12 +1136,28 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             else
             {
                 Debug.Assert(((object)conflict2 == null));
-                // https://github.com/dotnet/roslyn/issues/34453: We might need to do extra consistency check for events/properties and their 
-                //                                                accessors. Need to verify if what CheckForImplementationOfCorrespondingPropertyOrEvent
-                //                                                is doing should be applicable here as well.
             }
 
-            return implicitImpl;
+            return defaultImpl;
+
+            static bool stopLookup(MethodSymbol interfaceAccessor, TypeSymbol implementingType)
+            {
+                if (interfaceAccessor is null)
+                {
+                    return false;
+                }
+
+                SymbolAndDiagnostics symbolAndDiagnostics = implementingType.FindImplementationForInterfaceMemberInNonInterfaceWithDiagnostics(interfaceAccessor);
+
+                if (symbolAndDiagnostics.Symbol is object)
+                {
+                    return !symbolAndDiagnostics.Symbol.ContainingType.IsInterface;
+                }
+
+                // It is still possible that we actually looked for the accessor in interfaces, but failed due to an ambiguity.
+                // Let's try to look for a property to improve diagnostics in this scenario.
+                return !symbolAndDiagnostics.Diagnostics.Any(d => d.Code == (int)ErrorCode.ERR_MostSpecificImplementationIsNotFound);
+            }
         }
 
         private static Symbol FindMostSpecificImplementation(Symbol interfaceMember, NamedTypeSymbol implementingInterface, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
@@ -1062,6 +1167,16 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             switch (implementingMember.Count)
             {
                 case 0:
+                    (MethodSymbol interfaceAccessor1, MethodSymbol interfaceAccessor2) = GetImplementableAccessors(interfaceMember);
+
+                    // If interface actually implements an event or property accessor, but doesn't implement the event/property,
+                    // do not look for its implementation in bases.
+                    if ((interfaceAccessor1 is object && FindImplementationInInterface(interfaceAccessor1, implementingInterface).Count != 0) ||
+                        (interfaceAccessor2 is object && FindImplementationInInterface(interfaceAccessor2, implementingInterface).Count != 0))
+                    {
+                        return null;
+                    }
+
                     return FindMostSpecificImplementationInBases(interfaceMember, implementingInterface,
                                                                  ref useSiteDiagnostics,
                                                                  out var _, out var _);
@@ -1093,115 +1208,228 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             out Symbol conflictingImplementation1,
             out Symbol conflictingImplementation2)
         {
-            var implementations = ArrayBuilder<(MultiDictionary<Symbol, Symbol>.ValueSet MethodSet, MultiDictionary<NamedTypeSymbol, NamedTypeSymbol> Bases)>.GetInstance();
+            ImmutableArray<NamedTypeSymbol> allInterfaces = implementingType.AllInterfacesWithDefinitionUseSiteDiagnostics(ref useSiteDiagnostics);
 
-            foreach (var interfaceType in implementingType.AllInterfacesWithDefinitionUseSiteDiagnostics(ref useSiteDiagnostics))
+            if (allInterfaces.IsEmpty)
             {
-                if (!interfaceType.IsInterface)
+                conflictingImplementation1 = null;
+                conflictingImplementation2 = null;
+                return null;
+            }
+
+            // Properties or events can be implemented in an unconventional manner, i.e. implementing accessors might not be tied to a property/event.
+            // If we simply look for a more specific implementing property/event, we might find one with not most specific implementing accessors.
+            // Returning a property/event like that would be incorrect because runtime will use most specific accessor, or it will fail because there will
+            // be an ambiguity for the accessor implementation.
+            // So, for events and properties we look for most specific implementation of corresponding accessors and then try to tie them back to
+            // an event/property, if any.
+            (MethodSymbol interfaceAccessor1, MethodSymbol interfaceAccessor2) = GetImplementableAccessors(interfaceMember);
+
+            if (interfaceAccessor1 is null && interfaceAccessor2 is null)
+            {
+                return findMostSpecificImplementationInBases(interfaceMember, allInterfaces, ref useSiteDiagnostics, out conflictingImplementation1, out conflictingImplementation2);
+            }
+
+            Symbol accessorImpl1 = findMostSpecificImplementationInBases(interfaceAccessor1 ?? interfaceAccessor2, allInterfaces, ref useSiteDiagnostics,
+                                                                         out Symbol conflictingAccessorImplementation11, out Symbol conflictingAccessorImplementation12);
+
+            if (accessorImpl1 is null && conflictingAccessorImplementation11 is null) // implementation of accessor is not found
+            {
+                conflictingImplementation1 = null;
+                conflictingImplementation2 = null;
+                return null;
+            }
+
+            if (interfaceAccessor1 is null || interfaceAccessor2 is null)
+            {
+                if (accessorImpl1 is object)
                 {
-                    // this code is reachable in error situations
-                    continue;
+                    conflictingImplementation1 = null;
+                    conflictingImplementation2 = null;
+                    return findImplementationInInterface(interfaceMember, accessorImpl1);
                 }
 
-                MultiDictionary<Symbol, Symbol>.ValueSet candidate = FindImplementationInInterface(interfaceMember, interfaceType);
+                conflictingImplementation1 = findImplementationInInterface(interfaceMember, conflictingAccessorImplementation11);
+                conflictingImplementation2 = findImplementationInInterface(interfaceMember, conflictingAccessorImplementation12);
 
-                if (candidate.Count == 0)
+                if ((conflictingImplementation1 is null) != (conflictingImplementation2 is null))
                 {
-                    continue;
+                    conflictingImplementation1 = null;
+                    conflictingImplementation2 = null;
                 }
 
-                for (int i = 0; i < implementations.Count; i++)
-                {
-                    (MultiDictionary<Symbol, Symbol>.ValueSet methodSet, MultiDictionary<NamedTypeSymbol, NamedTypeSymbol> bases) = implementations[i];
-                    Symbol previous = methodSet.First();
-                    NamedTypeSymbol previousContainingType = previous.ContainingType;
+                return null;
+            }
 
-                    if (previousContainingType.Equals(interfaceType, TypeCompareKind.CLRSignatureCompareOptions))
+            Symbol accessorImpl2 = findMostSpecificImplementationInBases(interfaceAccessor2, allInterfaces, ref useSiteDiagnostics,
+                                                                         out Symbol conflictingAccessorImplementation21, out Symbol conflictingAccessorImplementation22);
+
+            if ((accessorImpl2 is null && conflictingAccessorImplementation21 is null) || // implementation of accessor is not found
+                (accessorImpl1 is null) != (accessorImpl2 is null)) // there is most specific implementation for one accessor and an ambiguous implementation for the other accessor. 
+            {
+                conflictingImplementation1 = null;
+                conflictingImplementation2 = null;
+                return null;
+            }
+
+            if (accessorImpl1 is object)
+            {
+                conflictingImplementation1 = null;
+                conflictingImplementation2 = null;
+                return findImplementationInInterface(interfaceMember, accessorImpl1, accessorImpl2);
+            }
+
+            conflictingImplementation1 = findImplementationInInterface(interfaceMember, conflictingAccessorImplementation11, conflictingAccessorImplementation21);
+            conflictingImplementation2 = findImplementationInInterface(interfaceMember, conflictingAccessorImplementation12, conflictingAccessorImplementation22);
+
+            if ((conflictingImplementation1 is null) != (conflictingImplementation2 is null))
+            {
+                // One pair of conflicting accessors can be tied to an event/property, but the other cannot be tied to an event/property.
+                // Dropping conflict information since it only affects diagnostic.
+                conflictingImplementation1 = null;
+                conflictingImplementation2 = null;
+            }
+
+            return null;
+
+            static Symbol findImplementationInInterface(Symbol interfaceMember, Symbol inplementingAccessor1, Symbol implementingAccessor2 = null)
+            {
+                NamedTypeSymbol implementingInterface = inplementingAccessor1.ContainingType;
+
+                if (implementingAccessor2 is object && !implementingInterface.Equals(implementingAccessor2.ContainingType, TypeCompareKind.ConsiderEverything))
+                {
+                    // Implementing accessors are from different types, they cannot be tied to the same event/property.
+                    return null;
+                }
+
+                MultiDictionary<Symbol, Symbol>.ValueSet implementingMember = FindImplementationInInterface(interfaceMember, implementingInterface);
+
+                switch (implementingMember.Count)
+                {
+                    case 1:
+                        return implementingMember.Single();
+                    default:
+                        return null;
+                }
+            }
+
+            static Symbol findMostSpecificImplementationInBases(
+                Symbol interfaceMember,
+                ImmutableArray<NamedTypeSymbol> allInterfaces,
+                ref HashSet<DiagnosticInfo> useSiteDiagnostics,
+                out Symbol conflictingImplementation1,
+                out Symbol conflictingImplementation2)
+            {
+                var implementations = ArrayBuilder<(MultiDictionary<Symbol, Symbol>.ValueSet MethodSet, MultiDictionary<NamedTypeSymbol, NamedTypeSymbol> Bases)>.GetInstance();
+
+                foreach (var interfaceType in allInterfaces)
+                {
+                    if (!interfaceType.IsInterface)
                     {
-                        // Last equivalent match wins
-                        implementations[i] = (candidate, bases);
-                        candidate = default;
-                        break;
+                        // this code is reachable in error situations
+                        continue;
                     }
 
-                    if (bases == null)
+                    MultiDictionary<Symbol, Symbol>.ValueSet candidate = FindImplementationInInterface(interfaceMember, interfaceType);
+
+                    if (candidate.Count == 0)
                     {
-                        Debug.Assert(implementations.Count == 1);
-                        bases = previousContainingType.InterfacesAndTheirBaseInterfacesWithDefinitionUseSiteDiagnostics(ref useSiteDiagnostics);
-                        implementations[i] = (methodSet, bases);
+                        continue;
                     }
 
-                    if (bases.ContainsKey(interfaceType))
+                    for (int i = 0; i < implementations.Count; i++)
                     {
-                        // Previous candidate is more specific
-                        candidate = default;
-                        break;
-                    }
-                }
+                        (MultiDictionary<Symbol, Symbol>.ValueSet methodSet, MultiDictionary<NamedTypeSymbol, NamedTypeSymbol> bases) = implementations[i];
+                        Symbol previous = methodSet.First();
+                        NamedTypeSymbol previousContainingType = previous.ContainingType;
 
-                if (candidate.Count == 0)
-                {
-                    continue;
-                }
-
-                if (implementations.Count != 0)
-                {
-                    MultiDictionary<NamedTypeSymbol, NamedTypeSymbol> bases = interfaceType.InterfacesAndTheirBaseInterfacesWithDefinitionUseSiteDiagnostics(ref useSiteDiagnostics);
-
-                    for (int i = implementations.Count - 1; i >= 0; i--)
-                    {
-                        if (bases.ContainsKey(implementations[i].MethodSet.First().ContainingType))
+                        if (previousContainingType.Equals(interfaceType, TypeCompareKind.CLRSignatureCompareOptions))
                         {
-                            // new candidate is more specific
-                            implementations.RemoveAt(i);
+                            // Last equivalent match wins
+                            implementations[i] = (candidate, bases);
+                            candidate = default;
+                            break;
+                        }
+
+                        if (bases == null)
+                        {
+                            Debug.Assert(implementations.Count == 1);
+                            bases = previousContainingType.InterfacesAndTheirBaseInterfacesWithDefinitionUseSiteDiagnostics(ref useSiteDiagnostics);
+                            implementations[i] = (methodSet, bases);
+                        }
+
+                        if (bases.ContainsKey(interfaceType))
+                        {
+                            // Previous candidate is more specific
+                            candidate = default;
+                            break;
                         }
                     }
 
-                    implementations.Add((candidate, bases));
-                }
-                else
-                {
-                    implementations.Add((candidate, null));
-                }
-            }
-
-            Symbol result;
-
-            switch (implementations.Count)
-            {
-                case 0:
-                    result = null;
-                    conflictingImplementation1 = null;
-                    conflictingImplementation2 = null;
-                    break;
-                case 1:
-                    MultiDictionary<Symbol, Symbol>.ValueSet methodSet = implementations[0].MethodSet;
-                    switch (methodSet.Count)
+                    if (candidate.Count == 0)
                     {
-                        case 1:
-                            result = methodSet.Single();
-                            if (result.IsAbstract)
-                            {
-                                result = null;
-                            }
-                            break;
-                        default:
-                            result = null;
-                            break;
+                        continue;
                     }
 
-                    conflictingImplementation1 = null;
-                    conflictingImplementation2 = null;
-                    break;
-                default:
-                    result = null;
-                    conflictingImplementation1 = implementations[0].MethodSet.First();
-                    conflictingImplementation2 = implementations[1].MethodSet.First();
-                    break;
-            }
+                    if (implementations.Count != 0)
+                    {
+                        MultiDictionary<NamedTypeSymbol, NamedTypeSymbol> bases = interfaceType.InterfacesAndTheirBaseInterfacesWithDefinitionUseSiteDiagnostics(ref useSiteDiagnostics);
 
-            implementations.Free();
-            return result;
+                        for (int i = implementations.Count - 1; i >= 0; i--)
+                        {
+                            if (bases.ContainsKey(implementations[i].MethodSet.First().ContainingType))
+                            {
+                                // new candidate is more specific
+                                implementations.RemoveAt(i);
+                            }
+                        }
+
+                        implementations.Add((candidate, bases));
+                    }
+                    else
+                    {
+                        implementations.Add((candidate, null));
+                    }
+                }
+
+                Symbol result;
+
+                switch (implementations.Count)
+                {
+                    case 0:
+                        result = null;
+                        conflictingImplementation1 = null;
+                        conflictingImplementation2 = null;
+                        break;
+                    case 1:
+                        MultiDictionary<Symbol, Symbol>.ValueSet methodSet = implementations[0].MethodSet;
+                        switch (methodSet.Count)
+                        {
+                            case 1:
+                                result = methodSet.Single();
+                                if (result.IsAbstract)
+                                {
+                                    result = null;
+                                }
+                                break;
+                            default:
+                                result = null;
+                                break;
+                        }
+
+                        conflictingImplementation1 = null;
+                        conflictingImplementation2 = null;
+                        break;
+                    default:
+                        result = null;
+                        conflictingImplementation1 = implementations[0].MethodSet.First();
+                        conflictingImplementation2 = implementations[1].MethodSet.First();
+                        break;
+                }
+
+                implementations.Free();
+                return result;
+            }
         }
 
         internal static MultiDictionary<Symbol, Symbol>.ValueSet FindImplementationInInterface(Symbol interfaceMember, NamedTypeSymbol interfaceType)
@@ -1227,26 +1455,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             return interfaceType.GetExplicitImplementationForInterfaceMember(interfaceMember);
         }
 
-        /// <summary>
-        /// Since dev11 didn't expose a symbol API, it had the luxury of being able to accept a base class's claim that 
-        /// it implements an interface.  Roslyn, on the other hand, needs to be able to point to an implementing symbol
-        /// for each interface member.
-        /// 
-        /// DevDiv #718115 was triggered by some unusual metadata in a Microsoft reference assembly (Silverlight System.Windows.dll).
-        /// The issue was that a type explicitly implemented the accessors of an interface event, but did not tie them together with
-        /// an event declaration.  To make matters worse, it declared its own protected event with the same name as the interface
-        /// event (presumably to back the explicit implementation).  As a result, when Roslyn was asked to find the implementing member
-        /// for the interface event, it found the protected event and reported an appropriate diagnostic.  Would it should have done
-        /// (and does do now) is recognize that no event associated with the accessors explicitly implementing the interface accessors
-        /// and returned null.
-        /// 
-        /// We resolved this issue by introducing a new step into the interface mapping algorithm: after failing to find an explicit
-        /// implementation in a type, but before searching for an implicit implementation in that type, check for an explicit implementation
-        /// of an associated accessor.  If there is such an implementation, then immediately return the associated property or event,
-        /// even if it is null.  That is, never attempt to find an implicit implementation for an interface property or event with an
-        /// explicitly implemented accessor.
-        /// </summary>
-        private static bool IsExplicitlyImplementedViaAccessors(Symbol interfaceMember, TypeSymbol currType, out Symbol implementingMember)
+        private static (MethodSymbol interfaceAccessor1, MethodSymbol interfaceAccessor2) GetImplementableAccessors(Symbol interfaceMember)
         {
             MethodSymbol interfaceAccessor1;
             MethodSymbol interfaceAccessor2;
@@ -1269,10 +1478,47 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                     }
                 default:
                     {
-                        implementingMember = null;
-                        return false;
+                        interfaceAccessor1 = null;
+                        interfaceAccessor2 = null;
+                        break;
                     }
             }
+
+            if (!interfaceAccessor1.IsImplementable())
+            {
+                interfaceAccessor1 = null;
+            }
+
+            if (!interfaceAccessor2.IsImplementable())
+            {
+                interfaceAccessor2 = null;
+            }
+
+            return (interfaceAccessor1, interfaceAccessor2);
+        }
+
+        /// <summary>
+        /// Since dev11 didn't expose a symbol API, it had the luxury of being able to accept a base class's claim that 
+        /// it implements an interface.  Roslyn, on the other hand, needs to be able to point to an implementing symbol
+        /// for each interface member.
+        /// 
+        /// DevDiv #718115 was triggered by some unusual metadata in a Microsoft reference assembly (Silverlight System.Windows.dll).
+        /// The issue was that a type explicitly implemented the accessors of an interface event, but did not tie them together with
+        /// an event declaration.  To make matters worse, it declared its own protected event with the same name as the interface
+        /// event (presumably to back the explicit implementation).  As a result, when Roslyn was asked to find the implementing member
+        /// for the interface event, it found the protected event and reported an appropriate diagnostic.  What it should have done
+        /// (and does do now) is recognize that no event associated with the accessors explicitly implementing the interface accessors
+        /// and returned null.
+        /// 
+        /// We resolved this issue by introducing a new step into the interface mapping algorithm: after failing to find an explicit
+        /// implementation in a type, but before searching for an implicit implementation in that type, check for an explicit implementation
+        /// of an associated accessor.  If there is such an implementation, then immediately return the associated property or event,
+        /// even if it is null.  That is, never attempt to find an implicit implementation for an interface property or event with an
+        /// explicitly implemented accessor.
+        /// </summary>
+        private static bool IsExplicitlyImplementedViaAccessors(Symbol interfaceMember, TypeSymbol currType, out Symbol implementingMember)
+        {
+            (MethodSymbol interfaceAccessor1, MethodSymbol interfaceAccessor2) = GetImplementableAccessors(interfaceMember);
 
             Symbol associated1;
             Symbol associated2;
@@ -1294,9 +1540,13 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                     {
                         implementingMember = null;
                     }
-
-                    return true;
                 }
+                else
+                {
+                    implementingMember = null;
+                }
+
+                return true;
             }
 
             implementingMember = null;
@@ -1335,14 +1585,27 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         /// property/event and we found an accessor, then the accessor we found is invalid, so clear it.
         /// </summary>
         private static void CheckForImplementationOfCorrespondingPropertyOrEvent(MethodSymbol interfaceMethod, TypeSymbol implementingType, bool implementingTypeIsFromSomeCompilation,
-                                                                                 ref Symbol implicitImpl, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+                                                                                 ref Symbol implicitImpl)
         {
+            Debug.Assert(!implementingType.IsInterfaceType());
             Debug.Assert(interfaceMethod.IsAccessor());
 
             Symbol associatedInterfacePropertyOrEvent = interfaceMethod.AssociatedSymbol;
-            Symbol implementingPropertyOrEvent = implementingType.FindImplementationForInterfaceMember(associatedInterfacePropertyOrEvent, ref useSiteDiagnostics); // NB: uses cache
+
+            // Do not make any adjustments based on presence of default interface implementation for the property or event.
+            // We don't want an addition of default interface implementation to change an error situation to success for
+            // scenarios where the default interface implementation wouldn't actually be used at runtime.
+            // When we find an implicit implementation candidate, we don't want to not discard it if we would discard it when
+            // default interface implementation was missing. Why would presence of default interface implementation suddenly
+            // make the candidate suiatable to implement the interface? Also, if we discard the candidate, we don't want default interface 
+            // implementation to take over later, since runtime might still use the discarded candidate.
+            // When we don't find any implicit implementation candidate, returning accessor of default interface implementation
+            // doesn't actually help much because we would find it anyway (it is implemented explicitly).
+            Symbol implementingPropertyOrEvent = implementingType.FindImplementationForInterfaceMemberInNonInterface(associatedInterfacePropertyOrEvent,
+                                                                                                                     ignoreImplementationInInterfacesIfResultIsNotReady: true); // NB: uses cache
+
             MethodSymbol correspondingImplementingAccessor = null;
-            if ((object)implementingPropertyOrEvent != null)
+            if ((object)implementingPropertyOrEvent != null && !implementingPropertyOrEvent.ContainingType.IsInterface)
             {
                 switch (interfaceMethod.MethodKind)
                 {
@@ -1402,42 +1665,40 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             }
         }
 
+        private static void ReportDefaultInterfaceImplementationMatchDiagnostics(Symbol interfaceMember, TypeSymbol implementingType, Symbol implicitImpl, DiagnosticBag diagnostics)
+        {
+            if (interfaceMember.Kind == SymbolKind.Method && implementingType.ContainingModule != implicitImpl.ContainingModule)
+            {
+                // The default implementation is coming from a different module, which means that we probably didn't check
+                // for the required runtime capability or language version
+
+                LanguageVersion requiredVersion = MessageID.IDS_DefaultInterfaceImplementation.RequiredVersion();
+                LanguageVersion? availableVersion = implementingType.DeclaringCompilation?.LanguageVersion;
+                if (requiredVersion > availableVersion)
+                {
+                    diagnostics.Add(ErrorCode.ERR_LanguageVersionDoesNotSupportDefaultInterfaceImplementationForMember,
+                                    GetInterfaceLocation(interfaceMember, implementingType),
+                                    implicitImpl, interfaceMember, implementingType,
+                                    MessageID.IDS_DefaultInterfaceImplementation.Localize(),
+                                    availableVersion.GetValueOrDefault().ToDisplayString(),
+                                    new CSharpRequiredLanguageVersion(requiredVersion));
+                }
+
+                if (!implementingType.ContainingAssembly.RuntimeSupportsDefaultInterfaceImplementation)
+                {
+                    diagnostics.Add(ErrorCode.ERR_RuntimeDoesNotSupportDefaultInterfaceImplementationForMember,
+                                    GetInterfaceLocation(interfaceMember, implementingType),
+                                    implicitImpl, interfaceMember, implementingType);
+                }
+            }
+        }
+
         /// <summary>
         /// These diagnostics are for members that do implicitly implement an interface member, but do so
         /// in an undesirable way.
         /// </summary>
         private static void ReportImplicitImplementationMatchDiagnostics(Symbol interfaceMember, TypeSymbol implementingType, Symbol implicitImpl, DiagnosticBag diagnostics)
         {
-            if (implicitImpl.ContainingType.IsInterface)
-            {
-                if (interfaceMember.Kind == SymbolKind.Method && implementingType.ContainingModule != implicitImpl.ContainingModule)
-                {
-                    // The default implementation is coming from a different module, which means that we probably didn't check
-                    // for the required runtime capability or language version
-
-                    LanguageVersion requiredVersion = MessageID.IDS_DefaultInterfaceImplementation.RequiredVersion();
-                    LanguageVersion? availableVersion = implementingType.DeclaringCompilation?.LanguageVersion;
-                    if (requiredVersion > availableVersion)
-                    {
-                        diagnostics.Add(ErrorCode.ERR_LanguageVersionDoesNotSupportDefaultInterfaceImplementationForMember,
-                                        GetInterfaceLocation(interfaceMember, implementingType),
-                                        implicitImpl, interfaceMember, implementingType,
-                                        MessageID.IDS_DefaultInterfaceImplementation.Localize(),
-                                        availableVersion.GetValueOrDefault().ToDisplayString(),
-                                        new CSharpRequiredLanguageVersion(requiredVersion));
-                    }
-
-                    if (!implementingType.ContainingAssembly.RuntimeSupportsDefaultInterfaceImplementation)
-                    {
-                        diagnostics.Add(ErrorCode.ERR_RuntimeDoesNotSupportDefaultInterfaceImplementationForMember,
-                                        GetInterfaceLocation(interfaceMember, implementingType),
-                                        implicitImpl, interfaceMember, implementingType);
-                    }
-                }
-
-                return;
-            }
-
             bool reportedAnError = false;
 
             if (interfaceMember.Kind == SymbolKind.Method)
