@@ -471,14 +471,13 @@ namespace Microsoft.CodeAnalysis
                 newMetadata = peReference.GetMetadataNoCopy();
 
                 // make sure basic structure of the PE image is valid:
-                var assemblyMetadata = newMetadata as AssemblyMetadata;
-                if (assemblyMetadata != null)
+                if (newMetadata is AssemblyMetadata assemblyMetadata)
                 {
-                    bool dummy = assemblyMetadata.IsValidAssembly();
+                    _ = assemblyMetadata.IsValidAssembly();
                 }
                 else
                 {
-                    bool dummy = ((ModuleMetadata)newMetadata).Module.IsLinkedModule;
+                    _ = ((ModuleMetadata)newMetadata).Module.IsLinkedModule;
                 }
             }
             catch (Exception e) when (e is BadImageFormatException || e is IOException)
@@ -522,6 +521,27 @@ namespace Microsoft.CodeAnalysis
 
             metadata = null;
             return false;
+        }
+
+        internal AssemblyMetadata GetAssemblyMetadata(PortableExecutableReference peReference, DiagnosticBag diagnostics)
+        {
+            var metadata = GetMetadata(peReference, MessageProvider, Location.None, diagnostics);
+            Debug.Assert(metadata != null || diagnostics.HasAnyErrors());
+
+            if (metadata == null)
+            {
+                return null;
+            }
+
+            // require the metadata to be a valid assembly metadata:
+            var assemblyMetadata = metadata as AssemblyMetadata;
+            if (assemblyMetadata?.IsValidAssembly() != true)
+            {
+                diagnostics.Add(MessageProvider.CreateDiagnostic(MessageProvider.ERR_MetadataFileNotAssembly, Location.None, peReference.Display));
+                return null;
+            }
+
+            return assemblyMetadata;
         }
 
         /// <summary>
