@@ -97,25 +97,25 @@ namespace Microsoft.CodeAnalysis.CSharp.Scripting.UnitTests
         }
 
         [Fact]
-        public void TestCreateScriptDelegate()
+        public async Task TestCreateScriptDelegate()
         {
             // create a delegate for the entire script
             var script = CSharpScript.Create("1 + 2");
             var fn = script.CreateDelegate();
 
             Assert.Equal(3, fn().Result);
-            Assert.ThrowsAsync<ArgumentException>("globals", () => fn(new object()));
+            await Assert.ThrowsAsync<ArgumentException>("globals", () => fn(new object()));
         }
 
         [Fact]
-        public void TestCreateScriptDelegateWithGlobals()
+        public async Task TestCreateScriptDelegateWithGlobals()
         {
             // create a delegate for the entire script
             var script = CSharpScript.Create<int>("X + Y", globalsType: typeof(Globals));
             var fn = script.CreateDelegate();
 
-            Assert.ThrowsAsync<ArgumentException>("globals", () => fn());
-            Assert.ThrowsAsync<ArgumentException>("globals", () => fn(new object()));
+            await Assert.ThrowsAsync<ArgumentException>("globals", () => fn());
+            await Assert.ThrowsAsync<ArgumentException>("globals", () => fn(new object()));
             Assert.Equal(4, fn(new Globals { X = 1, Y = 3 }).Result);
         }
 
@@ -178,9 +178,9 @@ F();");
         }
 
         [Fact(Skip = "https://github.com/dotnet/roslyn/issues/170")]
-        public void TestRunDynamicVoidScriptWithTerminatingSemicolon()
+        public async Task TestRunDynamicVoidScriptWithTerminatingSemicolon()
         {
-            var result = CSharpScript.RunAsync(@"
+            await CSharpScript.RunAsync(@"
 class SomeClass
 {
     public void Do()
@@ -193,9 +193,9 @@ d.Do();"
         }
 
         [Fact(Skip = "https://github.com/dotnet/roslyn/issues/170")]
-        public void TestRunDynamicVoidScriptWithoutTerminatingSemicolon()
+        public async Task TestRunDynamicVoidScriptWithoutTerminatingSemicolon()
         {
-            var result = CSharpScript.RunAsync(@"
+            await CSharpScript.RunAsync(@"
 class SomeClass
 {
     public void Do()
@@ -300,30 +300,30 @@ throw e;", globals: new ScriptTests());
         }
 
         [Fact]
-        public void TestRunCreatedScriptWithUnexpectedGlobals()
+        public async Task TestRunCreatedScriptWithUnexpectedGlobals()
         {
             var script = CSharpScript.Create("X + Y");
 
             // Global variables passed to a script without a global type
-            Assert.ThrowsAsync<ArgumentException>("globals", () => script.RunAsync(new Globals { X = 1, Y = 2 }));
+            await Assert.ThrowsAsync<ArgumentException>("globals", () => script.RunAsync(new Globals { X = 1, Y = 2 }));
         }
 
         [Fact]
-        public void TestRunCreatedScriptWithoutGlobals()
+        public async Task TestRunCreatedScriptWithoutGlobals()
         {
             var script = CSharpScript.Create("X + Y", globalsType: typeof(Globals));
 
             //  The script requires access to global variables but none were given
-            Assert.ThrowsAsync<ArgumentException>("globals", () => script.RunAsync());
+            await Assert.ThrowsAsync<ArgumentException>("globals", () => script.RunAsync());
         }
 
         [Fact]
-        public void TestRunCreatedScriptWithMismatchedGlobals()
+        public async Task TestRunCreatedScriptWithMismatchedGlobals()
         {
             var script = CSharpScript.Create("X + Y", globalsType: typeof(Globals));
 
             //  The globals of type 'System.Object' is not assignable to 'Microsoft.CodeAnalysis.CSharp.Scripting.Test.ScriptTests+Globals'
-            Assert.ThrowsAsync<ArgumentException>("globals", () => script.RunAsync(new object()));
+            await Assert.ThrowsAsync<ArgumentException>("globals", () => script.RunAsync(new object()));
         }
 
         [Fact]
@@ -354,7 +354,7 @@ throw e;", globals: new ScriptTests());
         [Fact]
         public async Task TestRepl()
         {
-            string[] submissions = new[]
+            var submissions = new[]
             {
                 "int x = 100;",
                 "int y = x * x;",
