@@ -1,6 +1,5 @@
 ﻿' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-Imports System.Threading.Tasks
 Imports Microsoft.CodeAnalysis.Test.Utilities
 Imports Roslyn.Test.Utilities
 
@@ -11,8 +10,6 @@ Namespace Microsoft.VisualStudio.LanguageServices.UnitTests.DebuggerIntelliSense
         Public Async Function QueryVariables() As Task
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Main(args As String())
         Dim bar = From x In "asdf"
@@ -24,7 +21,7 @@ End Module</Document>
                            </Project>
                        </Workspace>
             Using state = TestState.CreateVisualBasicTestState(text, False)
-                Await VerifyCompletionAndDotAfter("x", state)
+                Await state.VerifyCompletionAndDotAfter("x")
             End Using
         End Function
 
@@ -32,8 +29,6 @@ End Module</Document>
         Public Async Function EnteringMethod() As Task
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     [|Sub Main(args As String())|]
         Dim z = 4
@@ -42,8 +37,8 @@ End Module</Document>
                            </Project>
                        </Workspace>
             Using state = TestState.CreateVisualBasicTestState(text, False)
-                Await VerifyCompletionAndDotAfter("args", state)
-                Await VerifyCompletionAndDotAfter("z", state)
+                Await state.VerifyCompletionAndDotAfter("args")
+                Await state.VerifyCompletionAndDotAfter("z")
             End Using
         End Function
 
@@ -51,8 +46,6 @@ End Module</Document>
         Public Async Function ExitingMethod() As Task
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Main(args As String())
         Dim z = 4
@@ -61,8 +54,8 @@ End Module</Document>
                            </Project>
                        </Workspace>
             Using state = TestState.CreateVisualBasicTestState(text, False)
-                Await VerifyCompletionAndDotAfter("args", state)
-                Await VerifyCompletionAndDotAfter("z", state)
+                Await state.VerifyCompletionAndDotAfter("args")
+                Await state.VerifyCompletionAndDotAfter("z")
             End Using
         End Function
 
@@ -70,8 +63,6 @@ End Module</Document>
         Public Async Function SingleLineLambda() As Task
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Main(args As String())
         Dim z = [|Function(x) x + 5|]
@@ -81,7 +72,7 @@ End Module</Document>
                            </Project>
                        </Workspace>
             Using state = TestState.CreateVisualBasicTestState(text, False)
-                Await VerifyCompletionAndDotAfter("x", state)
+                Await state.VerifyCompletionAndDotAfter("x")
             End Using
         End Function
 
@@ -89,8 +80,6 @@ End Module</Document>
         Public Async Function MultiLineLambda() As Task
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Main(args As String())
         Dim z = [|Function(x)|]
@@ -102,7 +91,7 @@ End Module</Document>
                            </Project>
                        </Workspace>
             Using state = TestState.CreateVisualBasicTestState(text, False)
-                Await VerifyCompletionAndDotAfter("x", state)
+                Await state.VerifyCompletionAndDotAfter("x")
             End Using
         End Function
 
@@ -110,8 +99,6 @@ End Module</Document>
         Public Async Function LocalVariables() As Task
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Main(args As String())
         Dim bar as String = "boo"
@@ -122,9 +109,9 @@ End Module</Document>
                            </Project>
                        </Workspace>
             Using state = TestState.CreateVisualBasicTestState(text, False)
-                Await VerifyCompletionAndDotAfter("bar", state)
-                Await VerifyCompletionAndDotAfter("y", state)
-                Await VerifyCompletionAndDotAfter("z", state)
+                Await state.VerifyCompletionAndDotAfter("bar")
+                Await state.VerifyCompletionAndDotAfter("y")
+                Await state.VerifyCompletionAndDotAfter("z")
             End Using
         End Function
 
@@ -132,8 +119,6 @@ End Module</Document>
         Public Async Function CompletionAfterReturn() As Task
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Main(args As String())
         Dim bar as String = "boo"
@@ -144,11 +129,11 @@ End Module</Document>
                            </Project>
                        </Workspace>
             Using state = TestState.CreateVisualBasicTestState(text, True)
-                Await VerifyCompletionAndDotAfter("bar", state)
-                Await VerifyCompletionAndDotAfter("y", state)
-                Await VerifyCompletionAndDotAfter("z", state)
+                Await state.VerifyCompletionAndDotAfter("bar")
+                Await state.VerifyCompletionAndDotAfter("y")
+                Await state.VerifyCompletionAndDotAfter("z")
                 state.SendReturn()
-                Await VerifyCompletionAndDotAfter("y", state)
+                Await state.VerifyCompletionAndDotAfter("y")
             End Using
         End Function
 
@@ -156,8 +141,6 @@ End Module</Document>
         Public Async Sub TypeALineTenTimes()
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Main(args As String())
         Dim xx as String = "boo"
@@ -177,6 +160,9 @@ End Module</Document>
                     Await state.WaitForAsynchronousOperationsAsync()
                     Await state.AssertCompletionSession()
                     state.SendReturn()
+                    Await state.AssertNoCompletionSession()
+                    state.SendReturn()
+                    Assert.DoesNotContain("z", state.GetLineTextFromCaretPosition(), StringComparison.Ordinal)
                 Next
             End Using
         End Sub
@@ -185,8 +171,6 @@ End Module</Document>
         Public Async Sub SignatureHelpInParameterizedConstructor()
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Main(args As String())
         Dim xx as String = "boo"
@@ -207,8 +191,6 @@ End Module</Document>
         Public Async Sub SignatureHelpInMethodCall()
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Main(args As String())
         Dim xx as String = "boo"
@@ -229,8 +211,6 @@ End Module</Document>
         Public Async Sub SignatureHelpInGenericMethod()
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Self(Of T)(goo as T)
         Return goo
@@ -255,8 +235,6 @@ End Module</Document>
         Public Async Sub CompletionInExpression()
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Self(Of T)(goo as T)
         Return goo
@@ -282,8 +260,6 @@ End Module</Document>
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
                                <ProjectReference>ReferencedProject</ProjectReference>
-                               <Document>$$</Document>
-
                                <Document>Module Program
 
     Sub Main(args As String())
@@ -313,8 +289,6 @@ End Class
         Public Async Sub CompletionForGenericType()
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Self(Of T)(goo as T)
         Return goo
@@ -339,8 +313,6 @@ End Module</Document>
         Public Async Sub LocalsInForBlock()
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
-
                                <Document>Module Program
     Sub Main(args As String())
         Dim xx as String = "boo"
@@ -355,34 +327,17 @@ End Module</Document>
                            </Project>
                        </Workspace>
             Using state = TestState.CreateVisualBasicTestState(text, False)
-                Await VerifyCompletionAndDotAfter("q", state)
-                Await VerifyCompletionAndDotAfter("xx", state)
-                Await VerifyCompletionAndDotAfter("z", state)
+                Await state.VerifyCompletionAndDotAfter("q")
+                Await state.VerifyCompletionAndDotAfter("xx")
+                Await state.VerifyCompletionAndDotAfter("z")
             End Using
         End Sub
-
-        Private Async Function VerifyCompletionAndDotAfter(item As String, state As TestState) As Task
-            If state.IsImmediateWindow Then
-                state.SendTypeChars("?")
-            End If
-            state.SendTypeChars(item)
-            Await state.WaitForAsynchronousOperationsAsync()
-            Await state.AssertSelectedCompletionItem(item)
-            state.SendTab()
-            state.SendTypeChars(".")
-            Await state.WaitForAsynchronousOperationsAsync()
-            Await state.AssertCompletionSession()
-            For i As Integer = 0 To item.Length
-                state.SendBackspace()
-            Next
-        End Function
 
         <WorkItem(1044441, "http://vstfdevdiv:8080/DevDiv2/DevDiv/_workitems/edit/1044441")>
         <ConditionalWpfFact(GetType(x86)), Trait(Traits.Feature, Traits.Features.DebuggingIntelliSense)>
         Public Async Sub StoppedOnEndSub()
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
                                <Document>Module Program
     Sub Main(o as Integer)
     [|End Sub|]
@@ -390,7 +345,7 @@ End Module</Document>
                            </Project>
                        </Workspace>
             Using state = TestState.CreateVisualBasicTestState(text, False)
-                Await VerifyCompletionAndDotAfter("o", state)
+                Await state.VerifyCompletionAndDotAfter("o")
             End Using
         End Sub
 
@@ -399,7 +354,6 @@ End Module</Document>
         Public Async Sub StoppedOnEndProperty()
             Dim text = <Workspace>
                            <Project Language="Visual Basic" CommonReferences="true">
-                               <Document>$$</Document>
                                <Document>Class C
     Public Property x As Integer
         Get
@@ -412,7 +366,7 @@ End Class</Document>
                            </Project>
                        </Workspace>
             Using state = TestState.CreateVisualBasicTestState(text, False)
-                Await VerifyCompletionAndDotAfter("value", state)
+                Await state.VerifyCompletionAndDotAfter("value")
             End Using
         End Sub
     End Class
