@@ -160,18 +160,15 @@ class Class1
             VisualStudio.Editor.SendKeys("<s");
             VisualStudio.Editor.Verify.CompletionItemsExist("see", "seealso", "summary");
 
-            if (LegacyCompletionCondition.Instance.ShouldSkip)
+            // 🐛 Workaround for https://github.com/dotnet/roslyn/issues/33824
+            var completionItems = VisualStudio.Editor.GetCompletionItems();
+            var targetIndex = Array.IndexOf(completionItems, "see");
+            var currentIndex = Array.IndexOf(completionItems, VisualStudio.Editor.GetCurrentCompletionItem());
+            if (currentIndex != targetIndex)
             {
-                // 🐛 Workaround for https://github.com/dotnet/roslyn/issues/33824
-                var completionItems = VisualStudio.Editor.GetCompletionItems();
-                var targetIndex = Array.IndexOf(completionItems, "see");
-                var currentIndex = Array.IndexOf(completionItems, VisualStudio.Editor.GetCurrentCompletionItem());
-                if (currentIndex != targetIndex)
-                {
-                    var key = currentIndex < targetIndex ? VirtualKey.Down : VirtualKey.Up;
-                    var keys = Enumerable.Repeat(key, Math.Abs(currentIndex - targetIndex)).Cast<object>().ToArray();
-                    VisualStudio.Editor.SendKeys(keys);
-                }
+                var key = currentIndex < targetIndex ? VirtualKey.Down : VirtualKey.Up;
+                var keys = Enumerable.Repeat(key, Math.Abs(currentIndex - targetIndex)).Cast<object>().ToArray();
+                VisualStudio.Editor.SendKeys(keys);
             }
 
             VisualStudio.Editor.SendKeys(VirtualKey.Enter);
@@ -219,7 +216,7 @@ class Class1
         }
 
         // 🐛 The async completion controller in 16.0 Preview 4 fails to account for brace completion sessions.
-        [ConditionalWpfFact(typeof(LegacyCompletionCondition)), Trait(Traits.Feature, Traits.Features.Completion)]
+        [WpfFact(Skip = "https://github.com/dotnet/roslyn/issues/33825"), Trait(Traits.Feature, Traits.Features.Completion)]
         [WorkItem(33825, "https://github.com/dotnet/roslyn/issues/33825")]
         public void CompletionUsesTrackingPointsInTheFaceOfAutomaticBraceCompletion()
         {
@@ -270,10 +267,7 @@ class Class1
                 'M',
                 Shift(VirtualKey.Enter));
 
-            if (LegacyCompletionCondition.Instance.ShouldSkip)
-            {
-                // Async completion commits the item and inserts a blank line
-                VisualStudio.Editor.Verify.TextContains(@"
+            VisualStudio.Editor.Verify.TextContains(@"
 class Class1
 {
     void Main(string[] args)
@@ -283,19 +277,6 @@ $$
     }
 }",
 assertCaretPosition: true);
-            }
-            else
-            {
-                VisualStudio.Editor.Verify.TextContains(@"
-class Class1
-{
-    void Main(string[] args)
-    {
-        Main$$
-    }
-}",
-assertCaretPosition: true);
-            }
         }
 
         [WpfFact, Trait(Traits.Feature, Traits.Features.Completion)]
@@ -316,10 +297,7 @@ class Class1
                 'M',
                 Shift(VirtualKey.Enter));
 
-            if (LegacyCompletionCondition.Instance.ShouldSkip)
-            {
-                // Async completion commits the item (even in suggestion mode) and inserts a blank line
-                VisualStudio.Editor.Verify.TextContains(@"
+            VisualStudio.Editor.Verify.TextContains(@"
 class Class1
 {
     void Main(string[] args)
@@ -329,20 +307,7 @@ $$
     }
 }",
 assertCaretPosition: true);
-            }
-            else
-            {
-                VisualStudio.Editor.Verify.TextContains(@"
-class Class1
-{
-    void Main(string[] args)
-    {
-        M
-$$
-    }
-}",
-assertCaretPosition: true);
-            }
+
         }
 
         [WpfFact, Trait(Traits.Feature, Traits.Features.Completion)]
