@@ -95,8 +95,7 @@ w.Content = g;");
             VisualStudio.InteractiveWindow.SubmitText("b = null; w.Close(); w = null;");
         }
 
-        // This test is flaky when legacy completion is enabled
-        [ConditionalWpfFact(typeof(AsyncCompletionCondition))]
+        [WpfFact]
         public void TypingHelpDirectiveWorks()
         {
             VisualStudio.InteractiveWindow.ShowWindow(waitForPrompt: true);
