@@ -36,7 +36,10 @@ namespace Microsoft.CodeAnalysis.Host
         [return: NotNull]
         public TLanguageService GetRequiredService<TLanguageService>() where TLanguageService : ILanguageService
         {
-            var service = GetService<TLanguageService>();
+            // Producing a [MaybeNull]T value results in a warning like default(T).
+            // We are investigating a more complex design for nullable analysis to solve this. See:
+            // https://github.com/dotnet/roslyn/issues/38638
+            var service = GetService<TLanguageService>()!;
             if (service == null)
             {
                 throw new InvalidOperationException(string.Format(WorkspacesResources.Service_of_type_0_is_required_to_accomplish_the_task_but_is_not_available_from_the_workspace, typeof(TLanguageService)));
